vue/8차/p_01~05/server/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.13.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2021.1
sqlparse==0.4.1
